+++ packages/amplify-graphql-conversation-transformer/src/grapqhl-conversation-transformer.ts
@@ -16,6 +16,7 @@ import { ConversationPrepareHandler } from './transformer-steps/conversation-pre
 import { ConversationResolverGenerator } from './transformer-steps/conversation-resolver-generator';
 import { ConversationFieldHandler } from './transformer-steps/conversation-field-handler';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
+import * as semver from 'semver';
 
 /**
  * Configuration for the Conversation Directive
@@ -113,11 +114,23 @@ export class ConversationTransformer extends TransformerPluginBase {
   prepare = (ctx: TransformerPrepareStepContextProvider): void => {
     this.prepareHandler.prepare(ctx, this.directives);
   };
-}
 
-const validate = (config: ConversationDirectiveConfiguration, ctx: TransformerContextProvider): void => {
-  const { field } = config;
-  if (field.type.kind !== 'NamedType' || field.type.name.value !== 'ConversationMessage') {
-    throw new InvalidDirectiveError('@conversation return type must be ConversationMessage');
-  }
-};
+  validate = (): void => {
+    for (const directive of this.directives) {
+      if (directive.field.type.kind !== 'NamedType' || directive.field.type.name.value !== 'ConversationMessage') {
+        throw new InvalidDirectiveError('@conversation return type must be ConversationMessage');
+      }
+      if (directive.handler && directive.functionName) {
+        throw new InvalidDirectiveError("'functionName' and 'handler' are mutually exclusive");
+      }
+      if (directive.handler) {
+        const eventVersion = semver.coerce(directive.handler.eventVersion);
+        if (eventVersion?.major !== 1) {
+          throw new Error(
+            `Unsupported custom conversation handler. Expected eventVersion to match 1.x, received ${directive.handler.eventVersion}`,
+          );
+        }
+      }
+    }
+  };
+}
